+++ src/product/tool-definitions.js
@@ -12,6 +12,7 @@ import { textProductivityDefinitions } from './definitions/text-productivity.js'
 import { dataDeveloperDefinitions } from './definitions/data-developer.js';
 import { jsonToolsExtraDefinitions } from './definitions/json-tools-extra.js';
 import { cssGeneratorToolDefinitions } from './definitions/css-generator-tools.js';
+import { xmlAndIdToolDefinitions } from './definitions/xml-and-id-tools.js';
 import { advancedMathDefinitions } from './definitions/advanced-math.js';
 import { advancedFinanceDefinitions } from './definitions/advanced-finance.js';
 import { advancedHealthDefinitions } from './definitions/advanced-health.js';
@@ -1044,6 +1045,7 @@ const allToolDefinitions = Object.freeze({
     ...dataDeveloperDefinitions,
     ...jsonToolsExtraDefinitions,
     ...cssGeneratorToolDefinitions,
+    ...xmlAndIdToolDefinitions,
     ...advancedMathDefinitions,
     ...advancedFinanceDefinitions,
     ...advancedHealthDefinitions,